+++ Changelog.md
@@ -4,6 +4,8 @@ Language Features:
  * Allow accessing external functions via contract and interface names to obtain their selector.
  * Inline Assembly: Support literals ``true`` and ``false``.
  * Allow interfaces to inherit from other interfaces
+ * Allow gas and value to be set in external function calls using ``f{gas: 10000, value: 4 ether}()``.
+ * Allow specifying the ``salt`` for contract creations and thus the ``create2`` opcode using ``new C{salt: 0x1234, value: 1 ether}(arg1, arg2)``.
 
 
 Compiler Features: